Java Arrays - Java Multi-Dimensional Arrays
Multidimensional Arrays
A multidimensional array is an array of arrays.

Multidimensional arrays are useful when you want to store data as a tabular form, like a table with rows and columns.

To create a two-dimensional array, add each array within its own set of curly braces:

Example

myNumbers is now an array with two arrays as its elements.