Java Arrays - Java Multi-Dimensional Arrays
Access Elements
To access the elements of the myNumbers array, specify two indexes: one for the array, and one for the element inside that array. This example accesses the third element (2) in the second array (1) of myNumbers:

Example

Remember that: Array indexes start with 0: [0] is the first element. [1] is the second element, etc.