Java Method Overloading
Method Overloading

Instead of defining two methods that should do the same thing, it is better to overload one.

In the example below, we overload the plusMethod method to work for both int and double:

Note: Multiple methods can have the same name as long as the number and/or type of parameters are different.