Java Class Attributes
Java Class Attributes
In the previous chapter, we used the term "variable" for x in the example (as shown below). It is actually an attribute of the class. Or you could say that class attributes are variables within a class:

ExampleGet your own Java Server
Create a class called "Main" with two attributes: x and y:

Another term for class attributes is fields.