Modify Attributes
If you don't want the ability to override existing values, declare the attribute as final:

The final keyword is useful when you want a variable to always store the same value, like PI (3.14159...).

The final keyword is called a "modifier". You will learn more about these in the Java Modifiers Chapter.